Refactor module names for ansible-test runners and builders
