+++ galaxy_importer/ansible_test/runners/local_image.py
